Added --noreload option to server
Prevents zombie server when run from /etc/init.d/tinkerserver. The Flask
server by default runs a child process, which is the actual webserver,
so that the parent can monitor the server files and restart the child
when they change. This, however, interferes with start-stop-daemon's
ability to stop the server. This change fixes that behavior.